+++ src/fiuba/algo3/tp2/algopoly/model/casillero/AvanceDinamico.java
@@ -29,7 +29,7 @@ public class AvanceDinamico implements Encasillable {
 
         if (sumaDados == 7 || sumaDados == 8 || sumaDados == 9 || sumaDados == 10) {
 
-            jugador.mover(jugador.getCapital().getCantidad() % sumaDados);
+            jugador.mover( (int) jugador.getCapital().getCantidad() % sumaDados);
 
         }
